ml/requirements.txt
numpy
Pillow
opencv-python
scikit-image
tensorflow